# docs/swagger.yaml
basePath: /
definitions:
...
  request.PostCommentCreateRequest:
    properties:
      comment:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  request.PostCommentUpdateRequset:
    properties:
      comment:
        type: string
      id:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  request.PostCreateRequest:
    properties:
      attachments:
        items:
          type: string
        type: array
      content:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  request.PostUpdateRequset:
    properties:
      attachments:
        items:
          type: string
        type: array
      content:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
      sort_by:
        type: string
      total_pages:
        type: integer
      total_rows:
        type: integer
    type: object
  utils.ResponseData:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Pagination'
      status:
        type: string
    type: object
host: localhost:5001
info:
  contact:
    email: sarikhin@yahoo.co.id
    name: API Support
  description: This is a API Server for SIRNAWA
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIRNAWA API
  version: "1.0"
paths:
...
  /post:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List Post
      tags:
      - Post
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PostCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create Post
      tags:
      - Post
  /post-comment:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List PostComment
      tags:
      - PostComment
    post:
      consumes:
      - application/json
      parameters:
      - description: Create PostComment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PostCommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create PostComment
      tags:
      - PostComment
  /post-comment/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: PostComment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete PostComment By ID
      tags:
      - PostComment
    get:
      consumes:
      - application/json
      parameters:
      - description: PostComment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find PostComment By ID
      tags:
      - PostComment
    put:
      consumes:
      - application/json
      parameters:
      - description: Update PostComment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PostCommentUpdateRequset'
      - description: PostComment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update PostComment
      tags:
      - PostComment
  /post-comment/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated PostComment
      tags:
      - PostComment
  /post/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete Post By ID
      tags:
      - Post
    get:
      consumes:
      - application/json
      parameters:
      - description: Post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find Post By ID
      tags:
      - Post
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Post
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PostUpdateRequset'
      - description: Post id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update Post
      tags:
      - Post
  /post/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated Post
      tags:
      - Post
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"